audio: changed buffer offset and length types to size_t

+++ cpp/audio/in_memory_resource.cpp
@@ -15,10 +15,10 @@ InMemoryResource::InMemoryResource(category_t category, int id,
 	buffer = loader->get_resource();
 }
 
-std::tuple<const int16_t*,uint32_t> InMemoryResource::get_data(
-		uint32_t position, uint32_t data_length) {
+std::tuple<const int16_t*,size_t> InMemoryResource::get_data(
+		size_t position, size_t data_length) {
 	// if the resource's end has been reached
-	uint32_t length = static_cast<uint32_t>(buffer.size());
+	size_t length = buffer.size();
 	if (position >= length) {
 		return std::make_tuple(nullptr, 0);
 	}
